+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Use `MSBuildTasks` by default and when fail use `DotNetTasks` as fallback. (Rollback: #85)
 - Add `Tools.MSBuild` to support MSBuild from Visual Studio 2026. (Fix: #90)
 - Add `BuildExtension.BuildTools` with list of tools to build in order by `msbuild` and `dotnet`.
+- Add `BuildToolsExtension` to make simple to use `MSBuildOnly` and `dotnetBuildOnly` inside `INukeBuild`.
 
 ## [1.11.3] / 2026-04-16
 ### Updated